- wrapper script to run tests - flag for config sq_root_n_special -> O(N) recommendations
git-svn-id: https://moo.cmcl.cs.cmu.edu/svn/scaleron/src/java_sim@95 0bd769cf-3f23-0410-95ed-bb08bbc236b7

+++ run.bash
@@ -1,7 +1,13 @@
 #!/usr/bin/env bash
 
+if [ $# -lt 2 ]
+then
+        echo "Usage: ./run.bash neuron.properties numNodes"
+        exit
+fi
+
 export CLASSPATH=ext/mina-core-1.1.2.jar:ext/mina-filter-compression-1.1.2.jar:ext/slf4j-api-1.4.3.jar:ext/slf4j-simple-1.4.3.jar:.
 echo $CLASSPATH
-java -Dneuron.config=neuron.properties edu.cmu.neuron2.RonTest 10 & thepid=$!
+java -Dneuron.config=$1 edu.cmu.neuron2.RonTest $2 & thepid=$!
 sleep 60
 kill $!
